CodeOdessey/backend/requirements.txt
# Core Flask dependencies
Flask
Flask-CORS

# Audio processing
openai-whisper
librosa
sounddevice

# Scientific computing
scipy
numpy

# Build tools
setuptools
wheel
